Moved the interfaces to the "mapper" folder, and restructured the types for better organization.
