+++ src/mapper/helper/index.ts
@@ -1,6 +1,8 @@
+import * as ListForm from "./listForm";
 import * as SPConfig from "./spCfg";
 import * as WebPart from "./webpart";
 export {
+    ListForm,
     SPConfig,
     WebPart
 };
